requirements.txt
pycryptodome>=3.19.0
pyinstaller>=6.0.0
